Refactor funcs, extract to common
