+++ admin/cleanup-dev-env.ts
@@ -1,0 +1,3 @@
+#!/usr/bin/env -S deno --allow-run --allow-net
+// SPDX-License-Identifier: MIT
+import { runOrExit, parseProcessOutput } from "./procs.ts";
